Combine User and Group Config, to ensure that groups are always created before the users are
